Refactored the window component rendering
