+++ src/ui/ConnectToServer/index.tsx
@@ -107,7 +107,6 @@ class ConnectToServerContainer extends React.Component<
       }
       this.setLatestUrl(credentials.url);
       await main.showServerAdministration({
-        type: 'server-administration',
         credentials,
         validateCertificates: true,
       });
